refactor(ActivateUser): rename ActivateUser to ActivateFormOptions
